+++ tests/fixtures/noqa.py
@@ -561,4 +561,10 @@ unhashable = {[]}  # noqa: WPS443
 assert []  # noqa: WPS444
 unhashable = [] * 2  # noqa: WPS435
 
+def literal_none_func(arg: Literal[None]):  # noqa: WPS701
+    '''Literal[None]'''
+
+def literal_none_return_func() -> Literal[None]:  # noqa: WPS701
+    '''Literal[None]'''
+
 pi = 3.14 # noqa: WPS447