+++ CHANGELOG.md
@@ -10,6 +10,10 @@ We used to have incremental versioning before `0.1.0`.
 
 - Forbids using `Literal[None]` in function annotations
 
+### Bugfixes
+
+- Fixes **ImplicitElifViolation** false positives on a specific edge cases.
+
 
 ## 0.13.0