+++ CHANGELOG.md
@@ -4,10 +4,22 @@ We follow Semantic Versions since the `0.1.0` release.
 We used to have incremental versioning before `0.1.0`.
 
 
+## 0.13.0 WIP
+
+### Features
+
+- Forbids using `Literal[None]` in function annotations
+
+### Bugfixes
+
+- Fixes **ImplicitElifViolation** false positives on a specific edge cases.
+
+
 ## 0.13.0
 
 ### Features
 
+- Improves Github Action stability
 - `WPS431` now allow customize whitelist via `nested-classes-whitelist` setting
 - Forbids to have invalid strings like `**{'@': 1}`
 - Forbids to use implicit primitive values in a form of `lambda`